% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Annotation of the chromatographic peaks}
\usage{
annotatePeaks(dat_xcms, dat_chroma, ppmTol = 5, polarity = "positive",
  MCQindex = 0.5, excAddcutsForm = c("[M+Na]+", "[M+K]+"),
  assignGE = FALSE)
}
\arguments{
\item{dat_xcms}{XCMSnExp object}

\item{dat_chroma}{XChromatograms object}

\item{ppmTol}{Mass tolerance refers to the fraction of mass in parts-per-million (ppm) between the measured mass and theoretical mass  (default values: 5).The contaminant peaks are annotated based on the mass error (in ppm) between candidate peaks and contaminants in the open-source contaminants databases.}

\item{polarity}{The 'positive' or 'negative' polarity for LC-MS data}

\item{MCQindex}{This MCQ threshold selects high confidence chromatograms from low confidence chromatograms. If a chromatogram has a MCQ index lower than the threshold (default values: 0.5), the program will annotate it as low confidence and will remove the chromatogram from the peak table.}

\item{excAddcutsForm}{Peaks with specified adduct formation (e.g. [M+Na]+,[M+K]+) entered in the text box will be removed from the peak table.}

\item{assignGE}{Defines if result should be assigned to env. (TRUE) or not (FALSE).}
}
\value{
annotatePeaks objects
}
\description{
AnnotatePeaks provides annotation of the chromatographic peaks for the
non-targeted metabolomics.
}
\examples{
\dontrun{
anno <- annotatePeaks(dat_xcms, dat_chroma)
}
}
\keyword{annotatePeaks}
