% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{chromaList_cpva}
\alias{chromaList_cpva}
\title{chromaList_cpva}
\usage{
chromaList_cpva(x)
}
\arguments{
\item{x}{a object of featureChroma}
}
\description{
chromaList_cpva
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
