% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{chromaProfileMeric}
\alias{chromaProfileMeric}
\title{chromaProfileMeric}
\usage{
chromaProfileMeric(xy)
}
\arguments{
\item{xy}{a object of chromaProfileMeric}
}
\description{
chromaProfileMeric
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
