% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpvaOT.R
\name{cpvaOT}
\alias{cpvaOT}
\title{cpvaOT}
\usage{
cpvaOT(x, filename, type = "filterPeaks")
}
\arguments{
\item{x}{Annotatepeaks objects}

\item{filename}{base file name to save report, .tsv file will be appended to this name for the report}

\item{type}{"filterPeaks" or "all"}
}
\value{
A .tsv file
}
\description{
cpvaOutputTable
}
\examples{
\dontrun{
cpvaOT(x,"cpvaOutput")
}
}
\keyword{cpvaOutputTable}
