% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpvaWeb.R
\name{cpvaWeb}
\alias{cpvaWeb}
\title{Visualization of the chromatographic peaks}
\usage{
cpvaWeb(x, y, assignGE = FALSE)
}
\arguments{
\item{x}{annotatePeaks objects}

\item{y}{XChromatograms object}

\item{assignGE}{Defines if result should be assigned to env. (TRUE) or not (FALSE).}
}
\value{
interactive web pages
}
\description{
The cpvaWeb was used to build interactive web pages with Rshiny, and visualization of the chromatographic peaks.
}
\examples{
\dontrun{
cpvaWeb(anno, dat_chroma)
}
}
\keyword{annotatePeaks}
