% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{filterPeaks}
\alias{filterPeaks}
\title{filterPeaks}
\usage{
filterPeaks(x, isoremov = TRUE, adduremov = c("[M+Na]+", "[M+K]+"),
  contaremov = TRUE, chromaremov = TRUE)
}
\arguments{
\item{x}{a object of filterPeaks}

\item{isoremov}{xx}

\item{adduremov}{xx}

\item{contaremov}{xx}

\item{chromaremov}{xx}
}
\description{
filterPeaks
}
\examples{
\dontrun{
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
