% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{grectPlot}
\alias{grectPlot}
\title{grectPlot}
\usage{
grectPlot(x, y, size)
}
\arguments{
\item{XCMSnExp}{a object of peakgroup}
}
\description{
grectPlot
}
\examples{
\dontrun{
x
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
