% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{ionAnnotation}
\alias{ionAnnotation}
\title{ionAnnotation}
\usage{
ionAnnotation(object, polarity = "positive", annotation = TRUE,
  isotopeRemove = F)
}
\arguments{
\item{object}{xx}

\item{polarity}{xx}

\item{annotation}{xx}

\item{isotopeRemove}{xx}
}
\value{
ionAnnotation results
}
\description{
ionAnnotation
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{annotatePeaks}
