% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{merictable}
\alias{merictable}
\title{merictable}
\usage{
merictable(x)
}
\arguments{
\item{x}{a object of merictable}
}
\description{
merictable
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
