% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{peakgroup}
\alias{peakgroup}
\title{peakgroup}
\usage{
peakgroup(x)
}
\arguments{
\item{x}{a object of peakgroup}
}
\description{
peakgroup
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
