% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{peakRank}
\alias{peakRank}
\title{peakRank}
\usage{
peakRank(x, s, plotmargin = m)
}
\arguments{
\item{x}{a object of peakRank}

\item{s}{xx}

\item{plotmargin}{xx}
}
\description{
peakRank
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
