% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{pieplot_addu}
\alias{pieplot_addu}
\title{pieplot_addu}
\usage{
pieplot_addu(adducPie)
}
\arguments{
\item{adducPie}{a object of pieplot_addu}
}
\description{
pieplot_addu
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
