% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{pieplot_conta}
\alias{pieplot_conta}
\title{pieplot_conta}
\usage{
pieplot_conta(contaPie)
}
\arguments{
\item{contaPie}{a object of pieplot_conta}
}
\description{
pieplot_conta
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
