% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{pieplot_iso}
\alias{pieplot_iso}
\title{pieplot_iso}
\usage{
pieplot_iso(isoPie)
}
\arguments{
\item{isoPie}{a object of pieplot_iso}
}
\description{
pieplot_iso
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
