% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{readXCMSnExp}
\alias{readXCMSnExp}
\title{readXCMSnExp}
\usage{
readXCMSnExp(XCMSnExp)
}
\arguments{
\item{XCMSnExp}{a object of peakgroup}
}
\description{
readXCMSnExp
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
