% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{rectplot}
\alias{rectplot}
\title{rectplot}
\usage{
rectplot(x, y, size)
}
\arguments{
\item{x}{a object of rectplot}

\item{y}{a object of rectplot}

\item{size}{xx}
}
\description{
rectplot
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
