% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{shinyTuneChromatograms}
\alias{shinyTuneChromatograms}
\title{shinyTuneChromatograms}
\usage{
shinyTuneChromatograms(Chromatograms, MCQvalue)
}
\arguments{
\item{Chromatograms}{a object of pieplot_iso}

\item{MCQvalue}{xx}
}
\description{
shinyTuneChromatograms
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
