% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{shinyTuneXCMSnExp}
\alias{shinyTuneXCMSnExp}
\title{shinyTuneXCMSnExp}
\usage{
shinyTuneXCMSnExp(XCMSnExp, ppm, polarity)
}
\arguments{
\item{XCMSnExp}{XCMSnExp object}

\item{ppm}{ppm tollerence}

\item{polarity}{xx}
}
\description{
shinyTuneXCMSnExp provides to annotate the peaks
}
\examples{
\dontrun{
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
