% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_local.R
\name{xcms2peakTarget}
\alias{xcms2peakTarget}
\title{xcms2peakTarget}
\usage{
xcms2peakTarget(XCMSnExp)
}
\arguments{
\item{XCMSnExp}{a object of peakgroup}
}
\description{
xcms2peakTarget
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
